% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.st.R
\docType{data}
\name{data.st}
\alias{data.st}
\title{Macro (NUTS3) Panel}
\format{
A STFDF
}
\source{
\url{Authors' elaborations based on the Italian FADN data provided by CREA (https://www.crea.gov.it/). Authors' elaborations based on rainfall data taken from Funk et al. (2014, ftp://chg-ftpout.geog.ucsb.edu/pub/org/chg/products/CHIRPS-2.0/docs/USGS-DS832.CHIRPS.pdf).}
}
\usage{
data.st
}
\description{
A STDFDF object containing agricultural TFP relative levels and other attributes of NUTS3 regions in Italy. The data have been artificially altered to preserve confidentiality of information. Results may differ from those presented in the paper.
}
\keyword{datasets}
