% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distOte.R
\docType{data}
\name{distOte}
\alias{distOte}
\title{Matrix of technological distance}
\format{
A 106x106 matrix
}
\usage{
distOte
}
\description{
A matrix of agricultural technological distance of NUTS3. It is used to weight the geographical distance of NUTS3 regions.
}
\keyword{datasets}
