% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map1.R
\docType{data}
\name{map1}
\alias{map1}
\title{Cropland grid of Northen Italy (20km x 20km squares)}
\format{
SpatialPolygonsDataFrame object
}
\source{
\url{https://land.copernicus.eu/pan-european/corine-land-cover/clc-2012}
}
\usage{
map1
}
\description{
SpatialPolygonsDataFrame object with croplands of Northern Italy approximated with 20km x 20km squared polygons. It is based on croplands data contained in the Corine Land Cover 2012 raster map.
}
\keyword{datasets}
