% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2.R
\docType{data}
\name{map2}
\alias{map2}
\title{Cropland grid of Northen Italy (40km x 40km squares)}
\format{
SpatialPolygonsDataFrame object
}
\source{
\url{https://land.copernicus.eu/pan-european/corine-land-cover/clc-2012}
}
\usage{
map2
}
\description{
SpatialPolygonsDataFrame object with croplands of Northern Italy approximated with 40km x 40km squared polygons. It is based on croplands data contained in the Corine Land Cover 2012 raster map.
}
\keyword{datasets}
