% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map3.R
\docType{data}
\name{map3}
\alias{map3}
\title{Cropland grid of Northen Italy (60km x 60km squares)}
\format{
SpatialPolygonsDataFrame object
}
\source{
\url{https://land.copernicus.eu/pan-european/corine-land-cover/clc-2012}
}
\usage{
map3
}
\description{
SpatialPolygonsDataFrame object with croplands of Northern Italy approximated with 60km x 60km squared polygons. It is based on croplands data contained in the Corine Land Cover 2012 raster map.
}
\keyword{datasets}
