% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map4.R
\docType{data}
\name{map4}
\alias{map4}
\title{Cropland grid of Northen Italy (100km x 100km squares)}
\format{
SpatialPolygonsDataFrame object
}
\source{
\url{https://land.copernicus.eu/pan-european/corine-land-cover/clc-2012}
}
\usage{
map4
}
\description{
SpatialPolygonsDataFrame object with croplands of Northern Italy approximated with 100km x 100km squared polygons. It is based on croplands data contained in the Corine Land Cover 2012 raster map.
}
\keyword{datasets}
