% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r100km.R
\docType{data}
\name{r100km}
\alias{r100km}
\title{Cropland grid of Italy (100km x 100km squares)}
\format{
SpatialPolygonsDataFrame object
}
\source{
\url{https://land.copernicus.eu/pan-european/corine-land-cover/clc-2012}
}
\usage{
r100km
}
\description{
SpatialPolygonsDataFrame object with croplands of Italy approximated with 100km x 100km squared polygons. It is based on croplands data contained in the Corine Land Cover 2012 raster map.
}
\keyword{datasets}
