% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r20km.R
\docType{data}
\name{r20km}
\alias{r20km}
\title{Cropland grid of Italy (20km x 20km squares)}
\format{
SpatialPolygonsDataFrame object
}
\source{
\url{https://land.copernicus.eu/pan-european/corine-land-cover/clc-2012}
}
\usage{
r20km
}
\description{
SpatialPolygonsDataFrame object with croplands of Italy approximated with 20km x 20km squared polygons. It is based on croplands data contained in the Corine Land Cover 2012 raster map.
}
\keyword{datasets}
