% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r40km.R
\docType{data}
\name{r40km}
\alias{r40km}
\title{Cropland grid of Italy (40km x 40km squares)}
\format{
SpatialPolygonsDataFrame object
}
\source{
\url{https://land.copernicus.eu/pan-european/corine-land-cover/clc-2012}
}
\usage{
r40km
}
\description{
SpatialPolygonsDataFrame object with croplands of Italy approximated with 40km x 40km squared polygons. It is based on croplands data contained in the Corine Land Cover 2012 raster map.
}
\keyword{datasets}
