% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r60km.R
\docType{data}
\name{r60km}
\alias{r60km}
\title{Cropland grid of Italy (60km x 60km squares)}
\format{
SpatialPolygonsDataFrame object
}
\source{
\url{https://land.copernicus.eu/pan-european/corine-land-cover/clc-2012}
}
\usage{
r60km
}
\description{
SpatialPolygonsDataFrame object with croplands of Italy approximated with 60km x 60km squared polygons. It is based on croplands data contained in the Corine Land Cover 2012 raster map.
}
\keyword{datasets}
