% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsamp.R
\docType{data}
\name{regsamp}
\alias{regsamp}
\title{Coordinates of simulated farms in Northern Italy}
\format{
SpatialPointsDataFrame object
}
\usage{
regsamp
}
\description{
SpatialPointsDataFrame object with 1000 SpatialPoints to represent simulated farms in the simulation exercise
}
\keyword{datasets}
