% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabSard.R
\docType{data}
\name{tabSard}
\alias{tabSard}
\title{Neighbors of NUTS3 of Sardinia}
\format{
A data frame with the links
}
\usage{
tabSard
}
\description{
A table containing the link of NUTS3 in Sardinia with NUTS3 of mainland Italy
}
\keyword{datasets}
