% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.st.R
\docType{data}
\name{data.st}
\alias{data.st}
\title{Macro (NUTS3) Panel}
\format{
A STFDF
}
\usage{
data.st
}
\description{
A STDFDF object containing agricultural TFP relative levels and other attributes of NUTS3 regions in Italy. Data of NUTS3 regions with a low number of sampled farms (<15) have been artificially altered to preserve confidentiality.
}
\keyword{datasets}
