% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_regioni.R
\docType{data}
\name{sel_regioni}
\alias{sel_regioni}
\title{Merged NUTS3 for simulation exercise}
\format{
SpatialPolygons object
}
\source{
\url{https://www.istat.it/it/archivio/222527}
}
\usage{
sel_regioni
}
\description{
SpatialPolygons object with merged NUTS3 regions of Northern Italy used in the simulation exercise
}
\keyword{datasets}
